+++ module/utilities/welcome-message.js
@@ -1,0 +1,7 @@
+import { chatCardWelcomeMessage } from "./chat-cards.js";
+
+export function sendWelcomeMessage() {
+  ChatMessage.create({
+    content: chatCardWelcomeMessage()
+  })
+}
