+++ templates/actor/pc-sheet.html
@@ -11,7 +11,6 @@
   <header class="sheet-header">
     <div class="header-fields">
       <div class="flexrow">
-        <!-- <img class="profile-img" src="{{actor.img}}" data-edit="img" title="{{actor.name}}" /> -->
         <div class="profile-img-container">
           <img class="profile-img" src="{{actor.img}}" data-edit="img" title="{{actor.name}}" />
           {{#if data.isExclusiveTagActive}}
